+++ package.json
@@ -45,7 +45,7 @@
     "@types/react-router-dom": "^5.3.3",
     "autoprefixer": "^10.4.21",
     "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.10",
+    "tailwindcss": "^3.4.1",
     "typescript": "^4.9.5",
     "web-vitals": "^4.2.4"
   }
