+++ src/custom.d.ts
@@ -3,17 +3,3 @@ declare module "*.svg" {
   export default content;
 }
 
-export type Post = {
-  id: number;
-  title: string;
-  thumbnailUrl: string;
-  createdAt: string;
-  categories: string[];
-  content: string;
-}
-
-export type ContactForm = {
-  name: string;
-  email: string;
-  message: string;
-}
