+++ src/types/ContactForm.ts
@@ -1,0 +1,5 @@
+export type ContactForm = {
+  name: string;
+  email: string;
+  message: string;
+}
