+++ src/types/Post.ts
@@ -1,0 +1,9 @@
+export type Post = {
+  id: number;
+  title: string;
+  thumbnailUrl: string;
+  createdAt: string;
+  categories: string[];
+  content: string;
+}
+
